requirements.txt
numpy>=1.21.0
pandas>=1.5.0
torch>=2.0.0
yfinance>=0.2.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
streamlit>=1.22.0
plotly>=5.13.0
ta>=0.10.0  # Technical analysis library
tqdm>=4.62.3
pyyaml 